extract config.py from main.py
Why:

* wazo-agentd-wait is a lot faster without all the main.py imports
* homogenize with other daemons
