+++ wazo_agentd/config.py
@@ -1,0 +1,97 @@
+# Copyright 2012-2020 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+import argparse
+
+from xivo.chain_map import ChainMap
+from xivo.config_helper import (
+    read_config_file_hierarchy,
+    parse_config_file,
+)
+
+_DEFAULT_HTTP_PORT = 9493
+_DEFAULT_CONFIG = {
+    'user': 'wazo-agentd',
+    'debug': False,
+    'logfile': '/var/log/wazo-agentd.log',
+    'pidfile': '/run/wazo-agentd/wazo-agentd.pid',
+    'config_file': '/etc/wazo-agentd/config.yml',
+    'extra_config_files': '/etc/wazo-agentd/conf.d/',
+    'ami': {'host': 'localhost', 'username': 'wazo_agentd', 'password': 'die0Ahn8tae'},
+    'auth': {
+        'host': 'localhost',
+        'port': 9497,
+        'prefix': None,
+        'https': False,
+        'key_file': '/var/lib/wazo-auth-keys/wazo-agentd-key.yml',
+    },
+    'bus': {
+        'username': 'guest',
+        'password': 'guest',
+        'host': 'localhost',
+        'port': 5672,
+        'exchange_name': 'xivo',
+        'exchange_type': 'topic',
+    },
+    'consul': {
+        'scheme': 'http',
+        'host': 'localhost',
+        'port': 8500,
+        'verify': '/usr/share/xivo-certs/server.crt',
+    },
+    'rest_api': {
+        'listen': '127.0.0.1',
+        'port': _DEFAULT_HTTP_PORT,
+        'certificate': None,
+        'private_key': None,
+        'cors': {
+            'enabled': True,
+            'allow_headers': ['Content-Type', 'X-Auth-Token', 'Wazo-Tenant'],
+        },
+    },
+    'service_discovery': {
+        'enabled': True,
+        'advertise_address': 'localhost',
+        'advertise_port': _DEFAULT_HTTP_PORT,
+        'advertise_address_interface': 'eth0',
+        'refresh_interval': 25,
+        'retry_interval': 2,
+        'ttl_interval': 30,
+        'extra_tags': [],
+    },
+}
+
+
+def load(logger, argv):
+    cli_config = _parse_args()
+    file_config = read_config_file_hierarchy(ChainMap(cli_config, _DEFAULT_CONFIG))
+    service_key = _load_key_file(ChainMap(cli_config, file_config, _DEFAULT_CONFIG))
+    return ChainMap(cli_config, service_key, file_config, _DEFAULT_CONFIG)
+
+
+def _parse_args():
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        '-v', '--verbose', action='store_true', help='increase verbosity'
+    )
+    parser.add_argument('-u', '--user', action='store', help='User to run the daemon')
+
+    parsed = parser.parse_args()
+
+    config = {}
+    if parsed.verbose:
+        config['debug'] = parsed.verbose
+    if parsed.user:
+        config['user'] = parsed.user
+
+    return config
+
+
+def _load_key_file(config):
+    key_file = parse_config_file(config['auth']['key_file'])
+    return {
+        'auth': {
+            'username': key_file.get('service_id'),
+            'password': key_file.get('service_key'),
+        }
+    }
